+++ Changelog.md
@@ -3,6 +3,7 @@
 ## 1.7.1
 * NoNeighbourDelivery service added
 * NamedPersonOnly service added
+* SignedForByRecipient service added
 
 ## 1.7.0
 * dont allow Packstation and Postfiliale as delivery address street names for regular addresses